requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# LLM and API clients
anthropic>=0.25.0
httpx>=0.24.0

# Database drivers
psycopg2-binary>=2.9.0

# AWS services
boto3>=1.28.0

# Development and testing
faker>=19.0.0